# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SMS Blossom API
  version: 1.0.0
  description: >
    REST API for Shopify SMS marketing app (Express, ESM).
    Includes OAuth-driven backend, consent flows, discounts, reporting, webhooks, and GDPR.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sms-blossom.com
    description: Production API
...
security: []
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Get shop settings (rules, timezone)
      operationId: getSettings
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Settings payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettingsGetResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Settings]
      summary: Update shop settings (quiet hours, caps, abandoned delay)
      operationId: updateSettings
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SettingsPutRequest' }
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettingsGetResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /automations:
    get:
      tags: [Automations]
      summary: Read automations config
      operationId: getAutomations
      description: |
        Includes abandoned checkout settings:
        - `abandoned.delayMinutes`: inactivity window before SMS is sent (default 30)
        - `abandoned.discountCode`: optional code appended to recovery URL
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Automations config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationsResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Shop not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      tags: [Automations]
      summary: Update automations config
      operationId: updateAutomations
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AutomationsUpdateRequest' }
      responses:
        '200':
          description: Updated automations config
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationsResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Shop not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /segments:
    post:
      tags: [Segments]
      summary: Create or update a Segment
      operationId: createSegment
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: string }
                name: { type: string }
                filterJson: { type: object }
              required: [name, filterJson]
      responses:
        '200':
          description: Segment upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  segment: { type: object }
        '404':
          description: Unknown shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /segments/preview:
    post:
      tags: [Segments]
      summary: Preview segment audience (count + sample)
      operationId: previewSegment
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 25 }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                filterJson: { type: object }
      responses:
        '200':
          description: Count and sample returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  count: { type: integer }
                  sample: { type: array, items: { type: object } }
        '404':
          description: Unknown shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /reports/overview:
    get:
      tags: [Reports]
      summary: Overview KPIs
      operationId: getReportOverview
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: range
          schema: { type: string, enum: [7d, 30d, 90d], default: 30d }
      responses:
        '200':
          description: Overview metrics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReportOverviewResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Shop not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reports/campaigns:
    get:
      tags: [Reports]
      summary: Campaign attribution (revenue/orders via discount or UTM)
      operationId: getCampaignAttribution
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: window
          schema: { type: string, example: '30d' }
      responses:
        '200':
          description: Campaign attribution list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  range:
                    type: object
                    properties:
                      from: { type: string, format: date-time }
                      to: { type: string, format: date-time }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        campaignId: { type: string }
                        name: { type: string }
                        revenue: { type: number }
                        orders: { type: integer }
                        via:
                          type: object
                          properties:
                            discount: { type: integer }
                            utm: { type: integer }
                        clicks_lifetime: { type: integer }
                        messaging:
                          type: object
                          properties:
                            sent: { type: integer }
                            delivered: { type: integer }
                            failed: { type: integer }
        '404':
          description: Unknown shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /reports/automations:
    get:
      tags: [Reports]
      summary: Automation attribution (Abandoned Checkout revenue)
      operationId: getAutomationAttribution
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: window
          schema: { type: string, example: '30d' }
      responses:
        '200':
          description: Automation attribution
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  range:
                    type: object
                    properties:
                      from: { type: string, format: date-time }
                      to: { type: string, format: date-time }
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        automation: { type: string }
                        orders: { type: integer }
                        revenue: { type: number }
        '404':
          description: Unknown shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /reports/messaging/timeseries:
    get:
      tags: [Reports]
      summary: Messaging daily timeseries (sent/delivered/failed/cost)
      operationId: getMessagingTimeseries
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: window
          schema: { type: string, example: '7d' }
      responses:
        '200':
          description: Timeseries returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  range:
                    type: object
                    properties:
                      from: { type: string, format: date-time }
                      to: { type: string, format: date-time }
                  series:
                    type: array
                    items:
                      type: object
                      properties:
                        day: { type: string, format: date-time }
                        sent: { type: integer }
                        delivered: { type: integer }
                        failed: { type: integer }
                        cost: { type: number }
        '404':
          description: Unknown shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  schemas:
...
    SettingsGetResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        shop: { type: string, example: 'sms-blossom-dev.myshopify.com' }
        settings:
          type: object
          properties:
            timezone: { type: string, example: 'Europe/Athens' }
            quietHours:
              type: object
              properties:
                start: { type: integer, example: 21 }
                end: { type: integer, example: 9 }
            cap:
              type: object
              properties:
                windowHours: { type: integer, example: 24 }
                maxPerWindow: { type: integer, example: 1 }
            abandoned:
              type: object
              properties:
                delayMinutes: { type: integer, example: 30 }
    SettingsPutRequest:
      type: object
      properties:
        timezone: { type: string }
        quietHours:
          type: object
          properties:
            start: { type: integer }
            end: { type: integer }
          required: [start, end]
        cap:
          type: object
          properties:
            windowHours: { type: integer }
            maxPerWindow: { type: integer }
          required: [windowHours, maxPerWindow]
        abandoned:
          type: object
          properties:
            delayMinutes: { type: integer, minimum: 5, maximum: 1440 }
    ReportOverviewResponse:
      type: object
      properties:
        ok: { type: boolean }
        range: { type: string, example: '30d' }
        contacts:
          type: object
          properties:
            total: { type: integer }
            optedIn: { type: integer }
            optedOutRecent: { type: integer }
        messages:
          type: object
          properties:
            sent: { type: integer }
            delivered: { type: integer }
            failed: { type: integer }
            deliveryRate: { type: number, format: float }
        revenue:
          type: object
          properties:
            attributed: { type: number, format: float }
...
    AutomationRules:
      type: object
      properties:
        quietHours:
          type: object
          properties:
            enabled: { type: boolean }
            start: { type: integer, minimum: 0, maximum: 23 }
            end: { type: integer, minimum: 0, maximum: 23 }
            zone: { type: [string, 'null'] }
        frequencyCap:
          type: object
          properties:
            enabled: { type: boolean }
            per: { type: string, enum: [hour, day, week] }
            max: { type: integer, minimum: 1 }
        dedupeWindowMin: { type: integer, minimum: 0 }
    AutomationConfig:
      type: object
      properties:
        enabled: { type: boolean }
        template: { type: [string, 'null'] }
        delayMinutes: { type: integer, minimum: 5, maximum: 1440 }
        rules: { $ref: '#/components/schemas/AutomationRules' }
    AutomationsResponse:
      type: object
      properties:
        ok: { type: boolean }
        shop: { type: string }
        automations:
          type: object
          properties:
            orderPaid: { $ref: '#/components/schemas/AutomationConfig' }
            abandoned: { $ref: '#/components/schemas/AutomationConfig' }
            fulfillmentUpdate: { $ref: '#/components/schemas/AutomationConfig' }
            welcome: { $ref: '#/components/schemas/AutomationConfig' }
            backInStock: { $ref: '#/components/schemas/AutomationConfig' }
    AutomationsUpdateRequest:
      type: object
      properties:
        orderPaid: { $ref: '#/components/schemas/AutomationConfig' }
        abandoned: { $ref: '#/components/schemas/AutomationConfig' }
        fulfillmentUpdate: { $ref: '#/components/schemas/AutomationConfig' }
        welcome: { $ref: '#/components/schemas/AutomationConfig' }
        backInStock: { $ref: '#/components/schemas/AutomationConfig' }
      additionalProperties: false
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        details: {}